// Backend/NodeJS/microservices/user-service/src/user/entities/user.enums.ts
export enum UserRole {
  SYSTEM_MANAGER = 'SYSTEM_MANAGER',
  CLINIC_ADMIN = 'CLINIC_ADMIN',
  DOCTOR = 'DOCTOR',
  PATIENT = 'PATIENT',
  SECRETARY = 'SECRETARY',
}

export enum UserStatus {
  PENDING = 'PENDING',
  PENDING_ACTIVATION = 'PENDING_ACTIVATION',
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SUSPENDED = 'SUSPENDED',
  DELETED = 'DELETED',
}